+++ cogs/fractalthorns.py
@@ -11,7 +11,6 @@
 import asyncio
 import datetime as dt
 import logging
-import logging.handlers
 import random
 import re
 from io import BytesIO
@@ -465,8 +464,7 @@ class Fractalthorns(discord.Cog):
 			fractalthorns_api.CacheTypes.SKETCHES, ignore_stale=True
 		)
 
-		if sketches is None:
-			return {}
+		sketches.pop(None, None)
 
 		return list(sketches.keys())